+++ packages/coin-support/src/index.ts
@@ -4,6 +4,7 @@ import { CoinSupport } from '@cypherock/coin-support-interfaces';
 import { NearSupport } from '@cypherock/coin-support-near';
 import { SolanaSupport } from '@cypherock/coin-support-solana';
 import { TronSupport } from '@cypherock/coin-support-tron';
+import { XrpSupport } from '@cypherock/coin-support-xrp';
 import { coinFamiliesMap, CoinFamily } from '@cypherock/coins';
 
 const coinSupportMap: Record<CoinFamily, CoinSupport> = {
@@ -12,6 +13,7 @@ const coinSupportMap: Record<CoinFamily, CoinSupport> = {
   [coinFamiliesMap.near]: new NearSupport(),
   [coinFamiliesMap.solana]: new SolanaSupport(),
   [coinFamiliesMap.tron]: new TronSupport(),
+  [coinFamiliesMap.xrp]: new XrpSupport(),
 };
 
 export const getCoinSupport = (coinFamily: string) => {