+++ packages/cysync-core-constants/src/i18n/types.ts
@@ -1,4 +1,10 @@
-import { DeviceErrorCodes, IErrorMsg, ServerErrorType } from '../types';
+import {
+  DeviceErrorCodes,
+  IErrorMsg,
+  ServerErrorType,
+  ServerCoinErrorTypes,
+  CoinFamilyWithDefault,
+} from '../types';
 
 export interface LanguageStrings {
   x1Card: string;
@@ -276,15 +282,19 @@ interface LangSend {
       error: string;
       zeroAmount: string;
       notOverDustThreshold: string;
+      amountBelowXrpReserve: string;
+      balanceBelowXrpReserve: string;
     };
     fees: { title: string; label: string };
     warning: string;
     feeError: string;
+    feeBelowMinError: string;
     notEnoughBalance: string;
     toggleText: { replace: string; unconfirmed: string };
     infoBox: string;
     addButton: string;
     remarks: { label: string; placeholder: string; error: string };
+    destinationTag: { label: string; placeholder: string; error: string };
   };
   summary: {
     title: string;
@@ -294,6 +304,7 @@ interface LangSend {
     network: string;
     debit: string;
     remarks: string;
+    destinationTag: string;
   };
   finalMessage: {
     button: string;
@@ -349,6 +360,7 @@ interface LangHistory {
     description: string;
     feePrefix: { optimism: string };
     remarks: string;
+    destinationTag: string;
   };
   noData: { text: string; subText: string; buttonText: string };
   search: { placeholder: string; notFound: { text: string; subText: string } };
@@ -617,10 +629,13 @@ interface LangDeleteAccount {
   buttons: { yes: string; no: string };
 }
 
-interface LangErrors {
+export interface LangErrors {
   deviceErrors: Record<DeviceErrorCodes, IErrorMsg>;
   databaseError: IErrorMsg;
   serverErrors: Record<ServerErrorType, IErrorMsg>;
+  serverCoinErrors: Partial<
+    Record<CoinFamilyWithDefault, Record<ServerCoinErrorTypes, IErrorMsg>>
+  >;
   default: string;
   videoPlaybackError: string;
 }
@@ -884,7 +899,7 @@ interface LangDialogs {
         subTitle: string;
         actions: {
           confirmAuth: string;
-          enterPinAndTapCard: string;
+          tapCard: string;
         };
         messageBox: {
           warning: string;
@@ -1296,6 +1311,12 @@ interface LangInheritance {
         errorHeading: string;
         subtext: string;
       };
+    };
+  };
+  banner: {
+    title: string;
+    buttons: {
+      knowMore: string;
     };
   };
 }