+++ packages/cysync-core-constants/src/types/index.ts
@@ -1,2 +1,3 @@
 export * from './serverError';
 export * from './deviceError';
+export * from './serverCoinError';