+++ packages/cysync-core/scripts/dependencies/index.js
@@ -8,3 +8,4 @@ globalThis.ethers = require('ethers');
 globalThis.TronWeb = require('tronweb');
 globalThis.web3 = require('web3');
 globalThis.solanaWeb3 = require('@solana/web3.js');
+globalThis.xrpl = require('xrpl');