+++ packages/cysync-core/src/dialogs/Inheritance/PlanLogin/context/index.tsx
@@ -10,14 +10,8 @@ import React, {
 } from 'react';
 
 import { ITabs, useAsync, useMemoReturn, useTabsAndDialogs } from '~/hooks';
-import { InheritanceUserTypeMap, inheritancePlanService } from '~/services';
-import {
-  closeDialog,
-  IInheritancePlanDetails,
-  reminderPeriodStringToNumMap,
-  updateInheritancePlanDetails,
-  useAppDispatch,
-} from '~/store';
+import { InheritanceUserTypeMap } from '~/services';
+import { closeDialog, useAppDispatch } from '~/store';
 
 import {
   InheritancePlanLoginDialogContextInterface,
@@ -33,6 +27,7 @@ import {
   FetchRequestId,
   ValidateSignature,
 } from '../Dialogs';
+import { useUpdatePlanDetails } from '../../hooks/useUpdatePlanDetails';
 
 export const InheritancePlanLoginDialogContext: Context<InheritancePlanLoginDialogContextInterface> =
   createContext<InheritancePlanLoginDialogContextInterface>(
@@ -127,63 +122,11 @@ export const InheritancePlanLoginDialogProvider: FC<
     );
   }, [walletId]);
 
+  const { fetchAndUpdatePlan } = useUpdatePlanDetails({ authTokenConfig });
+
   const fetchPlanHanlder = useCallback(async () => {
     if (!authTokenConfig) return false;
-
-    const result = await inheritancePlanService.getPlan({
-      authTokenConfig,
-    });
-
-    if (result.error) {
-      throw result.error;
-    }
-
-    const activationDate =
-      result.result.subscription?.[0]?.activationDate ??
-      new Date().toISOString();
-
-    const expiryDate =
-      result.result.subscription?.[0]?.order?.expiryDate ??
-      Date.now() + 2 * 365 * 24 * 60 * 60 * 1000;
-
-    const planDetails: IInheritancePlanDetails = {
-      walletId: result.result.wallet,
-      name: result.result.fullName ?? '',
-      nominee:
-        result.result.nominee?.map(n => ({
-          name: n.name ?? '',
-          email: n.email ?? '',
-          alternateEmail: n.alternateEmail ?? '',
-        })) ?? [],
-      ...(result.result.executor !== undefined
-        ? {
-            executor: {
-              name: result.result.executor?.name ?? '',
-              email: result.result.executor?.email ?? '',
-              alternateEmail: result.result.executor?.alternateEmail ?? '',
-            },
-          }
-        : {}),
-      owner: {
-        name: result.result.owner?.name ?? '',
-        email: result.result.owner?.email ?? '',
-        alternateEmail: result.result.owner?.alternateEmail ?? '',
-      },
-      activationDate: new Date(activationDate).getTime(),
-      expiryDate: new Date(expiryDate).getTime(),
-      reminderPeriod:
-        reminderPeriodStringToNumMap[
-          result.result.emailConfig?.frequency ?? ''
-        ],
-    };
-
-    dispatch(
-      updateInheritancePlanDetails({
-        walletId: planDetails.walletId,
-        planDetails,
-      }),
-    );
-
+    await fetchAndUpdatePlan();
     onClose();
     return true;
   }, [authTokenConfig, onClose]);