+++ packages/cysync-core/src/dialogs/Inheritance/components/UserDetailsForm.tsx
@@ -1,3 +1,4 @@
+import { constants } from '@cypherock/cysync-core-constants';
 import {
   Container,
   Typography,
@@ -115,6 +116,7 @@ export const UserDetailsForm: FC<UserDetailsFormProps> = ({
             $letterSpacing: 'unset',
           }}
           disabled={isSubmittingUserDetails}
+          maxLength={constants.inheritance.maxUserDetailsInputLength}
         />
       </Container>
       <Container direction="row" $width="full" gap={24}>
@@ -134,6 +136,7 @@ export const UserDetailsForm: FC<UserDetailsFormProps> = ({
             $letterSpacing: 'unset',
           }}
           disabled={isSubmittingUserDetails}
+          maxLength={constants.inheritance.maxUserDetailsInputLength}
           $error={errorKey === 'email'}
         />
         <Input
@@ -155,6 +158,7 @@ export const UserDetailsForm: FC<UserDetailsFormProps> = ({
             $letterSpacing: 'unset',
           }}
           disabled={isSubmittingUserDetails}
+          maxLength={constants.inheritance.maxUserDetailsInputLength}
           $error={errorKey === 'alternateEmail'}
         />
       </Container>